% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADstat.hy.R
\name{ADstat.hy}
\alias{ADstat.hy}
\title{Provide the Anderson-Darling test statistic of standard normal distribution}
\usage{
ADstat.hy(x)
}
\arguments{
\item{x}{an input vector}
}
\description{
This function is based on ADStatQF function in MATLAB by Qing Feng.
}
