% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADstatWins.hy.R
\name{ADstatWins.hy}
\alias{ADstatWins.hy}
\title{Provide the winsorized Anderson-Darling test statistic}
\usage{
ADstatWins.hy(x)
}
\arguments{
\item{x}{an input vector}
}
\description{
This function winsorizes and calculates the A-D test statistic.
Based on the MATLAB function ADStatQF by Qing Feng.
}
