% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pileup.R
\name{annotate_pileup}
\alias{annotate_pileup}
\title{Annotating pileup}
\usage{
annotate_pileup(pileup, exon, inputType = "whole_intron",
  outputType = "part_intron")
}
\arguments{
\item{pileup}{raw coverage data}

\item{exon}{exon}

\item{inputType}{type of intronic region contained in pileup, with choices
"whole_intron", "part_intron", or "only_exon"; the first is the default.}

\item{outputType}{type of intronic region that will be included in output,
with choices "whole_intron", "part_intron", or "only_exon"; the second is
the default.}
}
\description{
This function is used to get pileup data based on new transcript for the
purpose of the downstream outlier detection analysis.
}
