% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{binarize}
\alias{binarize}
\title{binarize dataframe's column}
\usage{
binarize(inputData, column, operator, value)
}
\arguments{
\item{inputData}{data frame}

\item{column}{column to be function applied}

\item{operator}{">", ">=", "<", "<=", "==", "!=", "In", "Not In", "Contains", "Not Contains"}

\item{value}{value for used as criteria. in Case of In / Contains value should give as \code{c('EX1','EX2','EX3')} format}
}
\value{
data frame with changed column
}
\description{
change column as 1 or 0 with provided criteria
}
\examples{
iris \%>\% scissor::binarize('Sepal.Length', '>', '6')
iris \%>\% scissor::binarize('Species', 'In', 'c("virginica", "setosa")')

}
\seealso{
dplyr's \code{mutate_} function
}
