% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gaf.R
\name{build_gaf}
\alias{build_gaf}
\title{Build gene annotation}
\usage{
build_gaf(GTF.file = NULL, output.file = NULL)
}
\arguments{
\item{GTF.file}{GTF file name (with full directory). If NULL, UCSC annotation
databases will be used.}

\item{output.file}{an output file name where the gene annotation ranges will be saved}
}
\description{
Get gene annotation ranges (exons) for all genes from a GTF file.
}
\examples{
GTF.file = "genes.gtf"
regions = build_gaf(GTF.file=GTF.file)
write.table(x=regions,file="genes_gaf.txt",sep="\t",quote=F,row.names=F)

}
