% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_onoffout.R
\name{detect_onoffout}
\alias{detect_onoffout}
\title{Detect local shape outliers based on a residual matrix}
\usage{
detect_onoffout(resmat, rawmat, exonset, winlength = 100,
  readconstr = 10, siglev = NULL, cutoff = NULL)
}
\arguments{
\item{resmat}{a residual matrix from PCA, a data matrix subtracted by a low-rank matrix.}

\item{rawmat}{a raw coverage matrix, dataI from \link{process_data}}

\item{exonset}{exon/intron annotation matrix from \link{annotate_pileup}}

\item{winlength}{the window size. Default is 100.}

\item{readconstr}{the minimum reads count required to be considered for on/off local shape changes.
Default is 10.}
}
\description{
This function identifies local shape variants by mining locally on/off abnormalities.
}
