% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.cell.R
\name{evaluate.cell}
\alias{evaluate.cell}
\title{Cell-level evaluations on the Scissor selected cells}
\usage{
evaluate.cell(Load_file, Scissor_result, FDR_cutoff = 0.05, bootstrap_n = 100)
}
\arguments{
\item{Load_file}{File name for loading the preprocessed regression inputs.}

\item{Scissor_result}{Output variable from Scissor.}

\item{FDR_cutoff}{FDR cutoff in the correlation test for each pair of bulk sample and single-cell. The default value is 0.05.}

\item{bootstrap_n}{Bootstrap resampling times.}
}
\value{
This function returns a \code{data.frame} with rows are the Scissor selected cells. The columns contain the following attributes:
  \item{Mean correlation}{The mean correlation of a cell with all bulk samples.}
  \item{Correlation > 0}{The percentage of positive correlations of a cell with all bulk samples.}
  \item{Correlation < 0}{The percentage of negative correlations of a cell with all bulk samples.}
  \item{Significant Correlation}{The percentage of significant correlations (FDR < \code{FDR_cutoff}) of a cell with all bulk samples.}
  \item{Coefficient}{The regression coefficient beta calculated in Scissor.}
  \item{Beta 0\%}{The minimum value of bootstrap coefficient.}
  \item{Beta 25\%}{The lower quartile of bootstrap coefficient.}
  \item{Beta 50\%}{The median value of bootstrap coefficient.}
  \item{Beta 75\%}{TThe upper quartile of bootstrap coefficient.}
  \item{Beta 100\%}{The maximum value of bootstrap coefficient.}
  \item{Probability of zero}{The probability of a cell is not selected in bootstrap resamplings.}
}
\description{
This function perfroms evaluations for each Scissor selected cell.
}
\details{
This function uses two main strategies to
we used the nonparametric bootstrap strategy to assess the sampling distribution of coefficient or all the Scissor selected ddd cells.#'
}
