% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.exon.hy.R
\name{find.exon.hy}
\alias{find.exon.hy}
\title{Finding exons and introns}
\usage{
find.exon.hy(exon, is.intron = FALSE, num.intron = NULL)
}
\description{
This function is used to find exons and introns that will be used
in the downstream analysis
}
