% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscLocal.R
\name{get_offstat}
\alias{get_offstat}
\title{Quantifying abnormal deletions in narrow regions}
\usage{
get_offstat(residualData, pileupData, exonset, ADcutoff = 3,
  windowSize = 100, readconstr = 10)
}
\description{
This function quantifies abnormality associated with
}
