% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sparsePD4one.R
\name{get_sparsePD4one}
\alias{get_sparsePD4one}
\title{Determining the most outlying structure using sparse basis for one subject in a data set}
\usage{
get_sparsePD4one(basis, data, case)
}
\arguments{
\item{basis}{a matrix with sparse basis in columns}

\item{data}{normalized data through pre-processing procedure}

\item{case}{a subject ID number in the basis matrix}
}
\description{
This function discovers outlying structure for each subject as well as outlyingness scores
}
