% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splicedReads.R
\name{get_SRcount}
\alias{get_SRcount}
\title{Count splice junction reads from BAM files}
\usage{
get_SRcount(BAMfiles, caseIDs = NULL, regions = NULL, print.proc = FALSE)
}
\arguments{
\item{BAMfiles}{a list of BAM file names}

\item{caseIDs}{a list of case IDs in order of BAMfiles}

\item{regions}{genomic regions to read formatted as "chr1:1-100,200-300:+". See also \code{\link{build_gaf}}.}
}
\description{
Extract splice junction counts from BAM files for the provided regions.
}
\examples{
regions="chr1:1-100,200-300:+"
start_time = Sys.time()
countPileup = read_SRcount(BAMfiles=BAMfiles,caseIDs=caseIDs,regions=regions)
end_time = Sys.time()
end_time - start_time
}
