% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute data frame's data}
\usage{
impute(inputData, column, operator, value = NULL, to = NULL)
}
\arguments{
\item{inputData}{data frame}

\item{column}{column to be funcion applied}

\item{value}{keyword to remove or replace}

\item{to}{if operator is "Replace", target to be transformed}

\item{opeartor}{either 'Remove' or 'Replace'}
}
\value{
data frame with changed column
}
\description{
Remove or Replace specific value by given keyword
}
\examples{
impute(iris, 'Sepal.Width', 'Replace', value = '3.1', to = '3.12345')
impute(iris, 'Species', 'Remove', value = 'virginica')
}
\seealso{
dplyr's \code{filter_} and \code{mutate_} function
}
