% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len.intron.hy.R
\name{len.intron.hy}
\alias{len.intron.hy}
\title{Find intron length to be inserted}
\usage{
len.intron.hy(exon)
}
\description{
This function computes the average length of introns to be included in the SCISSOR analysis
}
