% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtransform_data.R
\name{logtransform_data}
\alias{logtransform_data}
\title{Transforming data}
\usage{
logtransform_data(
  inputData,
  logshiftVal = NULL,
  param.grid = NULL,
  draw.plot = FALSE
)
}
\arguments{
\item{inputData}{data to be transformed}
}
\description{
This function tansforms the input data
}
\keyword{log}
\keyword{parameter}
\keyword{shift}
