% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscGlobal.R
\name{miscGlobal}
\alias{miscGlobal}
\title{Detect global shape variants in RNA-seq inputData}
\usage{
miscGlobal(inputData, siglev = 1e-04, ADcutoff = 3, PCnum = NULL,
  maxPCnum = 10, reducedReturn = TRUE)
}
\arguments{
\item{inputData}{precessed RNA-seq data from \link{process_data}}

\item{siglev}{the significance level of the Chi-squared distribution. Default
is 1e-10.}

\item{ADcutoff}{a cutoff value for checking the normality based on
Anderson-Darling test statistic}

\item{PCnum}{the number of PCs to be used. If NULL (the default) the number
of PCs will be estimated.}

\item{maxPCnum}{the maximum number of PCs to be used. Default is 20.}

\item{reducedReturn}{logical, whether to show less results and not to return
large matrices. Default is TRUE.}

\item{subt.mean}{logical, whether to subtract mean before SVD. Default is
FALSE.}
}
\description{
This function discovers outlying subjects whose RNA-seq have abnormal shapes
and provides the most outlying direction for each outlier.
}
