% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscLocal.R
\name{miscLocal}
\alias{miscLocal}
\title{Detect local shape changes}
\usage{
miscLocal(miscGlobalResult, pileupData, Ranges, siglev = 1e-04,
  cutoff = NULL, ADcutoff = 3, windowSize = 100,
  reducedReturn = TRUE)
}
\arguments{
\item{miscGlobalResult}{Result from \link{miscGlobal}}

\item{pileupData}{raw coverage matrix, dataI from \link{process_data}}

\item{siglev}{the significance level for a robust outlier detection. Default is 1e-5.
IF cutoff is specified, siglev is not used.}

\item{cutoff}{the cutoff value for outlying statistics.
If NULL, the cutoff value is computed based on the specified siglev.}

\item{windowSize}{a window length. Default is 100.}

\item{reducedReturn}{}

\item{residualData}{a residual matrix from PCA, a data matrix subtracted by a low-rank matrix.}

\item{exonset}{data annotation matrix, exonset from \link{process_data}}
}
\description{
This function discovers outlying subjects whose RNA-seq have "local" abnormal shapes
and provides the most outlying window-direction for each outlier.
}
