% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_binarizeModule.R
\name{mod_binarizeModule_server}
\alias{mod_binarizeModule_server}
\title{Shiny Module for data transform}
\usage{
mod_binarizeModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
rounded data (not reactive)
}
\description{
binarizeModule Server Functions
}
