% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_binarizeModule.R
\name{mod_binarizeModule_ui}
\alias{mod_binarizeModule_ui}
\title{Shiny Module for data transform}
\usage{
mod_binarizeModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
binarizeModule UI Function
}
