% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_createModule.R
\name{mod_createModule_server}
\alias{mod_createModule_server}
\title{Shiny Module for feature creation}
\usage{
mod_createModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
createModule Server Functions
}
