% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_createModule.R
\name{mod_createModule_ui}
\alias{mod_createModule_ui}
\title{Shiny Module for feature create}
\usage{
mod_createModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
createModule UI Function
}
