% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_etcModule.R
\name{mod_etcModule_server}
\alias{mod_etcModule_server}
\title{Shiny Module for data transformation}
\usage{
mod_etcModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
etcModlue Server Functions
}
