% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exportModule.R
\name{mod_exportModule_server}
\alias{mod_exportModule_server}
\title{Shiny Module for data transform}
\usage{
mod_exportModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
rounded data (not reactive)
}
\description{
exportModule Server Functions
}
