% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exportModule.R
\name{mod_exportModule_ui}
\alias{mod_exportModule_ui}
\title{Shiny Module for data transform}
\usage{
mod_exportModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
exportModule UI Function
}
