% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gsubModule.R
\name{mod_gsubModule_server}
\alias{mod_gsubModule_server}
\title{Shiny Module for sub text replace}
\usage{
mod_gsubModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
gsubModule Server Functions
}
