% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_logModule.R
\name{mod_logModule_server}
\alias{mod_logModule_server}
\title{Shiny Module for data transform}
\usage{
mod_logModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data (not reactive)
}
\description{
logModule Server Functions
}
