% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reorderModule.R
\name{mod_reorderModule_server}
\alias{mod_reorderModule_server}
\title{Shiny Module for data transform}
\usage{
mod_reorderModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
rounded data (not reactive)
}
\description{
reorderModule Server Functions
}
