% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reorderModule.R
\name{mod_reorderModule_ui}
\alias{mod_reorderModule_ui}
\title{Shiny Module for data transform}
\usage{
mod_reorderModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
reorderModule UI Function
}
