% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_replaceModule.R
\name{mod_replaceModule_server}
\alias{mod_replaceModule_server}
\title{Shiny Module for data transformation}
\usage{
mod_replaceModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
transformed data
}
\description{
replaceModule Server Functions
}
