% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_replaceModule.R
\name{mod_replaceModule_ui}
\alias{mod_replaceModule_ui}
\title{Shiny Module for data transformation}
\usage{
mod_replaceModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
replaceModule UI Function
}
