% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roundModule.R
\name{mod_roundModule_server}
\alias{mod_roundModule_server}
\title{Shiny Module for data transform}
\usage{
mod_roundModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
rounded data (not reactive)
}
\description{
roundModule Server Functions
}
