% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_roundModule.R
\name{mod_roundModule_ui}
\alias{mod_roundModule_ui}
\title{Shiny Module for data transform}
\usage{
mod_roundModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
roundModule UI Function
}
