% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_splitModule.R
\name{mod_splitModule_server}
\alias{mod_splitModule_server}
\title{Shiny Module for data transform}
\usage{
mod_splitModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\value{
rounded data (not reactive)
}
\description{
splitModule Server Functions
}
