% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_splitModule.R
\name{mod_splitModule_ui}
\alias{mod_splitModule_ui}
\title{Shiny Module for data transform}
\usage{
mod_splitModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
splitModule UI Function
}
