% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCnum.gamma.hy.R
\name{PCnum.gamma.hy}
\alias{PCnum.gamma.hy}
\title{Estimate the number principal components for RNA-seq}
\usage{
PCnum.gamma.hy(eigenval, d, n, eps = 1e-04)
}
\arguments{
\item{eps}{tuning parameter for estimating the number of PCs. If NULL
(default), eps = (1/nrow(data))}
}
\description{
Estimate the number principal components for RNA-seq
}
