% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pileup.R
\name{plot_pileup}
\alias{plot_pileup}
\title{Plot pileup profiles}
\usage{
plot_pileup(Pileup, Ranges, cases = NULL, logcount = NULL,
  plot.meanpileup = TRUE, col.pileup = NULL, col.meanpileup = "grey",
  title = NULL, cex.title = 1.5, print.ranges = TRUE, xlim = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{Pileup}{a pileup object with rows for genomic positions and columns for
samples}

\item{Ranges}{a list of information for genomic regions.
\code{\link{get_Ranges}}.}

\item{cases}{a vector specifying which samples are to be plotted. Sample IDs
matching with the caseIDs in \code{\link{read_BAM}} or
\code{\link{read_pileup}} can be used. Also, the column numbers can be
simply used. e.g. \code{cases}=c(1,4,100) will plot the 1st, 4th, 100th
columns (i.e. samples) in the \code{Pileup}. If NULL, all samples with
plotted. Default is NULL.}

\item{logcount}{a numeric value indicating the number used in the logarithmic
transformation. When your data is the log-transformed one, this parameter
will print the raw read counts in y-axis labels instead of the log-scale.
For example, if your data is \code{Pileup} = log10(read depth+c)-log10(c),
"c" is the log count. Default is NULL, and y-axis will print labels based
on the scale of the \code{Pileup}. If your data is not log-scale, use NULL.}

\item{plot.meanpileup}{a logical value indicating whether the mean coverage
should be drawn. Default is TRUE.}

\item{col.pileup}{colors for the pileup samples}

\item{col.meanpileup}{a color for the mean coverage. Defult is "darkgrey"}

\item{print.ranges}{a logical value indicating whether genomic ranges and
locus ranges (\code{\link{get_Ranges}}) should be printed in x-axis.}

\item{...}{other graphical parameters passed to \code{points}.}
}
\description{
\code{plot_pileup} plots read counts of pileup coverage at base resolution.
}
\examples{
\\donotrun{
plot_pileup(Pileup=ScissorOutput$logData,Ranges=Ranges,cases=2,logcount=10)
}

}
\keyword{}
