% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ScissorOS.R
\name{plot_ScissorOS}
\alias{plot_ScissorOS}
\title{Plot outlyingness scores with a kernel density estimate}
\usage{
plot_ScissorOS(object, GSC = TRUE, LSC = TRUE, colmat = NULL,
  textSC = TRUE)
}
\arguments{
\item{object}{For the classes of GSCOutput and LSCOutput, the arguments GSC
and LSC will be ignored.}

\item{GSC}{logical, whether to plot the OS from GSC. Default is TRUE}

\item{LSC}{logical, whether to plot the OS from LSC. Default is TRUE}

\item{textSC}{}
}
\description{
Plot outlyingness scores with a kernel density estimate
}
