% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ScissorSC.R
\name{plot_ScissorSC}
\alias{plot_ScissorSC}
\title{Plot shape changes from SCISSOR}
\usage{
plot_ScissorSC(
  object,
  Data,
  SCids = NULL,
  GSC = TRUE,
  LSC = TRUE,
  subject.name = NULL,
  minfo = NULL,
  colmat = NULL
)
}
\description{
Plot shape changes from SCISSOR
}
