% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Processing data}
\usage{
process_data(pileup, exon, gene = NULL, inputType = "whole_intron",
  outputType = "part_intron", logshiftVal = NULL,
  plotNormalization = FALSE, ...)
}
\arguments{
\item{pileup}{a pile-up data matrix obtained from bam files, i.e. base-level
raw counts. Columns are samples and rows are genomic positions.}

\item{exon}{an exon annotation for the given gene.}

\item{inputType}{type of intronic region contained in pileup, with choices
"whole_intron", "part_intron", or "only_exon"; the second is the default.}

\item{outputType}{type of intronic region that will be included in output,
with choices "whole_intron", "part_intron", or "only_exon"; the second is
the default.}

\item{logshiftVal}{a pseudo count added to raw counts before the logarithmic
transformation. Default is NULL and automatically selects the pseudo count
based on the implemented algorithm.}

\item{plotNormalization}{}
}
\description{
Processing data
}
