% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_pileup}
\alias{process_pileup}
\title{Processing data}
\usage{
process_pileup(pileupData, Ranges, logshiftVal = NULL,
  plotNormalization = FALSE, ...)
}
\arguments{
\item{pileupData}{a pile-up data matrix obtained from bam files, i.e. base-level
raw counts. Columns are samples and rows are genomic positions.}

\item{Ranges}{an exon annotation for the given gene.}

\item{logshiftVal}{a pseudo count added to raw counts before the logarithmic
transformation. Default is NULL and automatically selects the pseudo count
based on the implemented algorithm.}

\item{plotNormalization}{}
}
\description{
Processing data
}
