% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BAM.R
\name{read_BAM}
\alias{read_BAM}
\title{Read BAM files into R}
\usage{
read_BAM(
  BAMfiles,
  caseIDs = NULL,
  gaf = NULL,
  symbol = NULL,
  regions = NULL,
  outputType = "part_intron",
  strand.specific = FALSE,
  ...
)
}
\arguments{
\item{BAMfiles}{a list of BAM file names}

\item{caseIDs}{a list of case IDs in order of BAMfiles}

\item{symbol}{a gene symbol (gene name)}

\item{regions}{genomic regions to read formatted as "chr1:1-100,200-300:+". See also \code{\link{build_gaf}}.}

\item{outputType}{type of intronic region that will be included in output,
with choices "whole_intron", "part_intron", or "only_exon"; the second is
the default.}

\item{strand.specific}{indicator to get only "+" strand pileup}
}
\description{
Get coverage pileup from BAM files for the provided regions.
}
\examples{
regions="chr1:1-100,200-300:+"
start_time = Sys.time()
countPileup = read_BAM(BAMfiles=BAMfiles,caseIDs=caseIDs,regions=regions)
end_time = Sys.time()
end_time - start_time

}
