% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pileup.R
\name{read_pileup}
\alias{read_pileup}
\title{Read coverage pileup into R}
\usage{
read_pileup(pileupfile, caseIDs = NULL, regions,
  inputType = "part_intron", outputType = "part_intron")
}
\arguments{
\item{regions}{genomic regions formatted as chr1:1-100,200-300:+"}

\item{inputType}{type of intronic region contained in pileup, with choices
"whole_intron", "part_intron", or "only_exon"; the first is the default.}

\item{outputType}{type of intronic region that will be included in output,
with choices "whole_intron", "part_intron", or "only_exon"; the second is
the default.}
}
\description{
This function is used to get pileup data based on new transcript for the
purpose of the downstream outlier detection analysis.
}
\examples{
package_path = "~/Dropbox/Research/hyochoi_GitHub/"
regions = "chrQ:7571719-7572198,7574858-7575157,7598088-7598437:-"
countPileup = read_pileup(pileupfile=paste0(package_path,"SCISSOR/toydata/TOY_coverage.txt"),inputType="whole_intron",caseIDs=NULL,regions=regions)

}
