% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.test.R
\name{reliability.test}
\alias{reliability.test}
\title{Reliability Significance Test}
\usage{
reliability.test(
  X,
  Y,
  network,
  alpha,
  family = c("gaussian", "binomial", "cox"),
  cell_num,
  n = 100,
  nfold = 10,
  verbose = TRUE
)
}
\arguments{
\item{X}{Scissor calculated correlation matrix for each pair of cells and bulk samples.}

\item{Y}{Scissor calculated response variable in the regression model.}

\item{network}{Scissor calculated cell-cell similarity network.}

\item{alpha}{Same parameter alpha used in Scissor.}

\item{family}{Same parameter family used in Scissor.}

\item{cell_num}{The number of the Scissor selected cells.}

\item{n}{Permutation times.}

\item{nfold}{The fold number in cross-validation.}
}
\value{
A list containing the following components:
  \item{statistic}{The test statistic.}
  \item{p}{The test p-value.}
  \item{Measurement_test_real}{The evaluation measurement on each fold calculated using the real label.}
  \item{Measurement_test_back}{A list with each component contains the evaluation measurements calculated using the permutated labels.}
}
\description{
Performs the reliability significance test to determine whether the inferred phenotype-to-cell
associations are reliable (statistical p-value less than 0.05) or are false positives.
}
\details{
This statistical test can determine whether the inferred phenotype-to-cell associations are
reliable (statistical p-value less than 0.05) or are false positives.

The motivation for the reliability significance test is: if the chosen single-cell and bulk data are not suitable for
the phenotype-to-cell associations, the correlations would be less informative and not well associated with the phenotype
labels. Thus, the corresponding prediction performance would be poor and not be significantly distinguishable from the
randomly permutated labels.

The evaluation measurements used in the reliability significance test are the mean squared error (MSE) for linear regression,
the area under the ROC curve (AUC) for classification, and the concordance index (c-index) for Cox regression.
}
