% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder column of Dataframe}
\usage{
reorder(inputData, columns)
}
\arguments{
\item{inputData}{data frame to reorder}

\item{columns}{vector of column names}
}
\value{
data frame with column order changed
}
\description{
select data as given column's order. Column provided as character.
}
\examples{
reorder(iris, c('Species', 'Sepal.Length', 'Sepal.Width', 'Petal.Length', 'Petal.Width'))
}
\seealso{
dplyr's \code{select_} function
}
