% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Scissor-package.R
\docType{package}
\name{Scissor-package}
\alias{Scissor}
\alias{Scissor-package}
\title{Scissor: Single-Cell Identification of Subpopulations with bulk Sample phenOtype coRrelation (SCISSOR)}
\description{
Scissor is a novel approach that utilizes the bulk phenotypes, such as disease stage, tumor metastasis, treatment response, and survival outcomes, to identify the most highly phenotype-associated cell subpopulations from single-cell data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sunduanchen/Scissor}
  \item Report bugs at \url{https://github.com/sunduanchen/Scissor/issues}
}

}
\author{
\strong{Maintainer}: Duanchen Sun \email{sund@ohsu.edu}

}
\keyword{internal}
