% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scissor.R
\name{Scissor}
\alias{Scissor}
\title{Scissor: Single-Cell Identification of Subsets with bulk RNA-Seq phenotype cORrelation}
\usage{
Scissor(
  alpha,
  family = c("binomial", "cox"),
  Given_Inputs = NULL,
  bulk_dataset,
  sc_dataset,
  survival_info = NULL,
  phenotype = NULL,
  cor_type = c("pearson", "spearman"),
  counts = TRUE,
  save_file = "Regression_Inputs.RData"
)
}
\arguments{
\item{alpha}{The parameter balances the effect of l1 norm and network-based penalties. The range of alpha is in \code{[0,1]}.
A larger alpha will lay more emphasis on l1 norm, i.e. Scissor will select fewer cells with more sparsity.}

\item{family}{Type of outcome. Can be binomial for discrete phenotype or cox for clinical outcome.}

\item{Given_Inputs}{Whether a saved regression input is given for tuning parameter. See tutorial for more details.}

\item{bulk_dataset}{The bulk RNA-seq expression data for Scissor. The rows are genes and the columns are samples.}

\item{sc_dataset}{The single-cell RNA-seq expression data for Scissor. The rows are genes and the columns are samples.}

\item{survival_info}{A three-column format of clinical data used for cox regression. The first column contains the
sample ID (observations). The second column is the survival time and the third column stands for the event
(e.g.recurrence of cancer or death), which is indicated by the variable "Status" with 0 (no event, censored) and 1 (event).}

\item{phenotype}{The discrete vector for logistic regression.}

\item{cor_type}{The correlation method in constructing the correlation matrix between bulk samples and single cells.}

\item{counts}{Whether the input single-cell data is count quantification. If yes, an additional normalization step is
applied in preprocessing steps.}

\item{save_file}{The file name for saving. Scissor can save the preprocessed regression inputs into a R data, which
can help users to directly tune model parameters. See tutorial for more details.}
}
\value{
This function will return a list with the following components:
  \item{score}{The vector of final MarkRank scores for each gene.}
  \item{steps}{The final iteration steps in random walk based scoring procedure.}
  \item{NET2}{The weighted adjacent matrix of gene cooperation network.}
  \item{initial_pars}{The initial/input parameter values used in MarkRank.}
  \item{dis}{The pairwise distance matrix of input network. This variable will be \code{Null} if input d=Inf.}
}
\description{
Scissor is a novel proposed network-based model, which can identify the cooperative
biomarkers for heterogeneous complex disease diagnoses.
}
\details{
Scissor is a novel proposed network-based model, which can identify the cooperative
biomarkers for heterogeneous complex disease diagnoses.
}
\references{
Duanchen Sun and Zheng Xia (2020): Scissor: Phenotype guided single-cell
subset identification from bulk RNA-seq.
}
