% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split}
\alias{split}
\title{Split data frame's column text}
\usage{
split(inputData, column, splitby)
}
\arguments{
\item{inputData}{data frame to split}

\item{column}{column that split will be applied}

\item{splitby}{keyword to used as split; ex. '.'}
}
\value{
data frame with newly splited columns
}
\description{
apply column split text and provides 2 result column. Column provided as character
}
\examples{
split(iris, column = 'Sepal.Length', splitby = '.')
}
\seealso{
tidyr's \code{separate} function
}
