% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{subset dataframe's column}
\usage{
subset(inputData, column, operator, value)
}
\arguments{
\item{inputData}{data frame}

\item{column}{column to be function applied}

\item{operator}{"In", "Not In", "Contains", "Not Contains", ">", ">=", "<", "<=", "==", "!="}

\item{value}{value for used as criteria. in Case of In / Contains value should give as \code{c('EX1','EX2','EX3')} format}
}
\value{
data frame with changed column
}
\description{
subset data with given criteria
}
\examples{
scissor::subset(iris,'Sepal.Length', '>', '5')
scissor::subset(iris,'Species', 'In', 'c("virginica", "setosa")')

}
\seealso{
dplyr's \code{filter_} function, scissor's \code{binarize} function.
}
