% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{transform (numerically) dataframe's column}
\usage{
trans(inputData, column, operator, value = NULL)
}
\arguments{
\item{inputData}{data frame}

\item{column}{column to be function applied (should be numerical column)}

\item{operator}{"round", "Log", "Log10", "Log2", "Sqrt", "-", "Min-Max", "Normal"}

\item{value}{value for used as additional parameter (round only for now)}
}
\value{
data frame with changed column
}
\description{
apply column with round, log, normalize.
}
\examples{
scissor::trans(iris,'Sepal.Length', '-')
scissor::trans(iris,'Sepal.Length', 'round')

}
\seealso{
dplyr's \code{mutate_} function
}
