% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gaf.R
\name{build_gaf}
\alias{build_gaf}
\title{Build gene annotation}
\usage{
build_gaf(Gene, GTF.file = NULL, hg.ref = c("hg19", "hg38"))
}
\arguments{
\item{Gene}{character(1) identifying the gene symbol, e.g. \code{"TP53"}}

\item{GTF.file}{GTF file name (with full directory). If NULL, UCSC annotation
databases will be used.}

\item{hg.ref}{either "hg19" or "hg38", which will be used for UCSC annotation
databases. By default, hg19 is used.}
}
\description{
Get genome ranges (exons) for a given gene from UCSC or a given GTF file.
}
\examples{
build_gaf(Gene="TP53")

}
