% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gaf.R
\name{build_gaf}
\alias{build_gaf}
\title{Build gene annotation}
\usage{
build_gaf(Gene, GTF.file = NULL)
}
\arguments{
\item{Gene}{character(1) identifying the gene symbol for gene list, e.g. \code{"TP53"}}

\item{GTF.file}{GTF file name (with full directory). If NULL, UCSC annotation
databases will be used.}
}
\description{
Get genome ranges (exons) for a given gene from a GTF file.
}
\examples{
GTF.file = "genes.gtf"
regions = build_gaf(Gene=genelist[1:100],GTF.file=GTF.file)
write.table(x=regions,file="genes_gaf.txt",sep="\t",quote=F,col.names=F,row.names=F)

}
