% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalizing log data}
\usage{
normalize_data(inputData, pileupData, Ranges, smoothness = 0.7,
  makePlot = FALSE, ...)
}
\arguments{
\item{inputData}{log-transformed data matrix (e.g. object from \link{logtransform_data})}

\item{pileupData}{raw coverage data matrix}

\item{makePlot}{logical, whether to plot variation with respect to overall expression with fitted curve.
Default is FALSE}

\item{exonset}{exon/intron annotation matrix from \link{process_data}}

\item{Gene}{gene name to print in the plot}
}
\description{
This function reduces the variation associated with sequentcing depths
remaining after a log transformation. The output is the main data object for the downstream
shape change detection \link{miscGlobal}.
}
