% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scissor.R
\name{Scissor}
\alias{Scissor}
\title{Scissor: Single-Cell Identification of Subpopulations with bulk Sample phenOtype coRrelation}
\usage{
Scissor(
  bulk_dataset,
  sc_dataset,
  phenotype,
  tag = NULL,
  alpha = NULL,
  cutoff = 0.2,
  family = c("gaussian", "binomial", "cox"),
  Save_file = "Scissor_inputs.RData",
  Load_file = NULL
)
}
\arguments{
\item{bulk_dataset}{Bulk expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{sc_dataset}{Single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample.
A Seurat object that contains the preprocessed data and constructed network is preferred. Otherwise, a cell-cell similarity network is
constructed based on the input matrix.}

\item{phenotype}{Phenotype annotation of each bulk sample. It can be a continuous dependent variable,
binary group indicator vector, or clinical survival data:
  \itemize{
  \item Continuous dependent variable. Should be a quantitative vector for \code{family = gaussian}.
  \item Binary group indicator vector. Should be either a 0-1 encoded vector or a factor with two levels for \code{family = binomial}.
  \item Clinical survival data. Should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable,
  with '1' indicating event (e.g.recurrence of cancer or death), and '0' indicating right censored.
  The function \code{Surv()} in package survival produces such a matrix.
  }}

\item{tag}{Names for each phenotypic group. Used for linear and logistic regressions only.}

\item{alpha}{Parameter used to balance the effect of the l1 norm and the network-based penalties. It can be a number or a searching vector.
If \code{alpha = NULL}, a default searching vector is used. The range of alpha is in \code{[0,1]}. A larger alpha lays more emphasis on the l1 norm.}

\item{cutoff}{Cutoff for the percentage of the Scissor selected cells in total cells. This parameter is used to restrict the number of the
Scissor selected cells. A cutoff less than \code{50\%} (default \code{20\%}) is recommended depending on the input data.}

\item{family}{Response type for the regression model. It depends on the type of the given phenotype and
can be \code{family = gaussian} for linear regression, \code{family = binomial} for classification, or \code{family = cox} for Cox regression.}

\item{Save_file}{File name for saving the preprocessed regression inputs into a RData.}

\item{Load_file}{File name for loading the preprocessed regression inputs. It can help to tune the model parameter \code{alpha}.
Please see Scissor Tutorial for more details.}
}
\value{
This function returns a list with the following components:
  \item{Coefs}{The regression coefficient for each cell.}
  \item{Scissor_pos}{The cell IDs of Scissor+ cells.}
  \item{Scissor_neg}{The cell IDs of Scissor- cells.}
}
\description{
\code{Scissor} is a novel approach that utilizes the phenotypes, such as disease stage, tumor metastasis, treatment response, and survival
outcomes, collected from bulk assays to identify the most highly phenotype-associated cell subpopulations from single-cell data.
}
\details{
Scissor is a novel algorithm to identify cell subpopulations from single-cell data that are most highly associated with the given phenotypes.
The three data sources for Scissor inputs are a single-cell expression matrix, a bulk expression matrix, and a phenotype of interest.
The phenotype annotation of each bulk sample can be a continuous dependent variable, binary group indicator vector, or clinical survival data.
The key step of Scissor is to quantify the similarity between the single-cell and bulk samples by Pearson correlation for each pair of cells and bulk samples.
After this, Scissor optimizes a regression model on the correlation matrix with the sample phenotype.
The selection of the regression model depends on the type of the input phenotype, i.e., linear regression for continuous variables,
logistic regression for dichotomous variables, and Cox regression for clinical survival data.
Based on the signs of the estimated regression coefficients, the cells with non-zero coefficients can be indicated as
Scissor positive (Scissor+) cells and Scissor negative (Scissor-) cells, which are positively and negatively associated
with the phenotype of interest, respectively.
}
\references{
Duanchen Sun and Zheng Xia (2021): Phenotype-guided subpopulation identification from single-cell sequencing data.
}
