% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scissor.R
\name{Scissor.v5.optimized}
\alias{Scissor.v5.optimized}
\title{Optimized Scissor Algorithm for Seurat ver5}
\usage{
Scissor.v5.optimized(
  bulk_dataset,
  sc_dataset,
  phenotype,
  tag = NULL,
  alpha = NULL,
  cutoff = 0.2,
  family = c("gaussian", "binomial", "cox"),
  Save_file = "Scissor_inputs.RData",
  Load_file = NULL,
  ...
)
}
\arguments{
\item{bulk_dataset}{Bulk expression matrix of related disease. Each row represents a gene and each column represents a sample.}

\item{sc_dataset}{Single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample.
A Seurat object that contains the preprocessed data and constructed network is preferred. Otherwise, a cell-cell similarity network is
constructed based on the input matrix.}

\item{phenotype}{Phenotype annotation of each bulk sample. It can be a continuous dependent variable,
binary group indicator vector, or clinical survival data:
  \itemize{
  \item Continuous dependent variable. Should be a quantitative vector for \code{family = gaussian}.
  \item Binary group indicator vector. Should be either a 0-1 encoded vector or a factor with two levels for \code{family = binomial}.
  \item Clinical survival data. Should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable,
  with '1' indicating event (e.g.recurrence of cancer or death), and '0' indicating right censored.
  The function \code{Surv()} in package survival produces such a matrix.
  }}

\item{tag}{Names for each phenotypic group. Used for linear and logistic regressions only.}

\item{alpha}{Parameter used to balance the effect of the l1 norm and the network-based penalties. It can be a number or a searching vector.
If \code{alpha = NULL}, a default searching vector is used. The range of alpha is in \code{[0,1]}. A larger alpha lays more emphasis on the l1 norm.}

\item{cutoff}{Cutoff for the percentage of the Scissor selected cells in total cells. This parameter is used to restrict the number of the
Scissor selected cells. A cutoff less than \code{50\%} (default \code{20\%}) is recommended depending on the input data.}

\item{family}{Response type for the regression model. It depends on the type of the given phenotype and
can be \code{family = gaussian} for linear regression, \code{family = binomial} for classification, or \code{family = cox} for Cox regression.}

\item{Save_file}{File name for saving the preprocessed regression inputs into a RData.}

\item{Load_file}{File name for loading the preprocessed regression inputs. It can help to tune the model parameter \code{alpha}.}

\item{...}{print `verbose` message, use `seed` for reproducibility and other optional arguments. Use `assay` to specify the assay slot of the Seurat object.}
}
\description{
Scissor.v5 from `https://doi.org/10.1038/s41587-021-01091-3`and `https://github.com/sunduanchen/Scissor/issues/59`
Another version of Scissor.v5() to optimize memory usage and execution speed in preprocess.
}
\section{LICENSE}{

Licensed under the GNU General Public License version 3 (GPL-3.0).
A copy of the license is available at <https://www.gnu.org/licenses/gpl-3.0.en.html>.
}

\references{
Sun D, Guan X, Moran AE, Wu LY, Qian DZ, Schedin P, et al. Identifying phenotype-associated subpopulations by integrating bulk and single-cell sequencing data. Nat Biotechnol. 2022 Apr;40(4):527–38.
}
\concept{scissor}
